feat: move dashboard to a separate directory
- Move dashboard files to a separate directory
- Updated test cases where needed

fix: LEARNER-9715
