+++ app/src/main/java/org/openedx/app/AppRouter.kt
@@ -24,7 +24,7 @@ import org.openedx.course.presentation.section.CourseSectionFragment
 import org.openedx.course.presentation.unit.container.CourseUnitContainerFragment
 import org.openedx.course.presentation.unit.video.VideoFullScreenFragment
 import org.openedx.course.presentation.unit.video.YoutubeVideoFullScreenFragment
-import org.openedx.dashboard.presentation.DashboardRouter
+import org.openedx.dashboard.presentation.dashboard.DashboardRouter
 import org.openedx.discovery.presentation.DiscoveryRouter
 import org.openedx.discovery.presentation.NativeDiscoveryFragment
 import org.openedx.discovery.presentation.search.CourseSearchFragment
