+++ dashboard/src/main/java/org/openedx/dashboard/presentation/dashboard/DashboardAnalytics.kt
@@ -1,4 +1,4 @@
-package org.openedx.dashboard.presentation
+package org.openedx.dashboard.presentation.dashboard
 
 interface DashboardAnalytics {
     fun dashboardCourseClickedEvent(courseId: String, courseName: String)
