+++ dashboard/src/test/java/org/openedx/dashboard/presentation/DashboardViewModelTest.kt
@@ -31,6 +31,9 @@ import org.openedx.core.system.notifier.AppUpgradeNotifier
 import org.openedx.core.system.notifier.CourseDashboardUpdate
 import org.openedx.core.system.notifier.CourseNotifier
 import org.openedx.dashboard.domain.interactor.DashboardInteractor
+import org.openedx.dashboard.presentation.dashboard.DashboardAnalytics
+import org.openedx.dashboard.presentation.dashboard.DashboardUIState
+import org.openedx.dashboard.presentation.dashboard.DashboardViewModel
 import java.net.UnknownHostException
 
 @OptIn(ExperimentalCoroutinesApi::class)
